refactor(state): split state.rs into per-type module files
Break the 1531-line state.rs into a state/ module of focused files, one
per type. The full impl State stays whole in mod.rs; the supporting data
types move to dedicated files and tests are split per concern next to the
file each group exercises.

- mod.rs:      State struct + entire impl State + re-exports
- tracked.rs:  Tracked
- account.rs:  AccountInfo, Account
- storage.rs:  StorageOverlay
- changes.rs:  StateChanges, StorageChangeSet
- journal.rs:  JournalEntry, StateCheckpoint

No behavior change; public re-exports from evm/mod.rs are unchanged.
