+++ crates/evm2/src/evm/state/journal.rs
@@ -1,0 +1,189 @@
+//! Revert journal and checkpoint types.
+
+use super::{Account, StorageOverlay};
+use crate::interpreter::Word;
+use alloy_primitives::Address;
+
+/// State checkpoint for reverting state changes.
+#[allow(missing_copy_implementations)]
+#[derive(Debug, Eq, PartialEq)]
+pub struct StateCheckpoint {
+    /// Revert journal length at the checkpoint.
+    pub(super) journal_len: usize,
+    /// Emitted log count at the checkpoint.
+    pub(super) logs_len: usize,
+}
+
+/// Compact journal entry for reverting state changes.
+#[derive(Clone, Debug, Eq, PartialEq)]
+#[non_exhaustive]
+pub enum JournalEntry {
+    /// Account current value changed.
+    AccountChange {
+        /// Account address.
+        address: Address,
+        /// Previous current account value.
+        previous: Option<Account>,
+    },
+    /// Account overlay entry was inserted.
+    AccountInserted {
+        /// Account address.
+        address: Address,
+    },
+    /// Account was touched.
+    Touch {
+        /// Account address.
+        address: Address,
+    },
+    /// Account was self-destructed.
+    SelfDestruct {
+        /// Account address.
+        address: Address,
+    },
+    /// Persistent storage changed.
+    StorageChange {
+        /// Account address.
+        address: Address,
+        /// Storage key.
+        key: Word,
+        /// Previous current storage value.
+        previous: Word,
+    },
+    /// Persistent storage slot overlay was inserted.
+    StorageInserted {
+        /// Account address.
+        address: Address,
+        /// Storage key.
+        key: Word,
+    },
+    /// Account storage wipe flag changed.
+    StorageWipe {
+        /// Account address.
+        address: Address,
+        /// Previous storage overlay.
+        previous: Option<StorageOverlay>,
+    },
+    /// Transient storage changed.
+    TransientStorageChange {
+        /// Account address.
+        address: Address,
+        /// Storage key.
+        key: Word,
+        /// Previous transient storage value.
+        previous: Option<Word>,
+    },
+    /// Account was warmed by EIP-2929 access tracking.
+    AccountWarmed {
+        /// Account address.
+        address: Address,
+    },
+    /// Storage slot was warmed by EIP-2929 access tracking.
+    StorageWarmed {
+        /// Account address.
+        address: Address,
+        /// Storage key.
+        key: Word,
+    },
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{
+        SpecId, Version,
+        evm::{
+            CacheDB,
+            state::{AccountInfo, State},
+        },
+    };
+    use alloc::vec::Vec;
+    use alloy_primitives::Log;
+
+    #[test]
+    fn destruct_change_rolls_back_to_checkpoint() {
+        let address = Address::from([0x33; 20]);
+        let mut state = State::new(CacheDB::default());
+
+        let checkpoint = state.checkpoint();
+        state.mark_destructed(&address);
+
+        assert!(state.is_selfdestructed(&address));
+        state.rollback(checkpoint, Version::base(SpecId::FRONTIER).features);
+        assert!(!state.is_selfdestructed(&address));
+    }
+
+    #[test]
+    fn log_rolls_back_to_checkpoint() {
+        use alloy_primitives::{Bytes, LogData};
+
+        let mut state = State::new(CacheDB::default());
+        let kept = Log {
+            address: Address::from([0x44; 20]),
+            data: LogData::new_unchecked(Vec::new(), Bytes::from_static(&[0x01])),
+        };
+        let reverted = Log {
+            address: Address::from([0x55; 20]),
+            data: LogData::new_unchecked(Vec::new(), Bytes::from_static(&[0x02])),
+        };
+
+        state.log(kept.clone());
+        let checkpoint = state.checkpoint();
+        state.log(reverted);
+
+        assert_eq!(
+            state.logs(),
+            &[
+                kept.clone(),
+                Log {
+                    address: Address::from([0x55; 20]),
+                    data: LogData::new_unchecked(Vec::new(), Bytes::from_static(&[0x02])),
+                }
+            ]
+        );
+        state.rollback(checkpoint, Version::base(SpecId::FRONTIER).features);
+        assert_eq!(state.logs(), &[kept]);
+    }
+
+    #[test]
+    fn spurious_dragon_rollback_preserves_precompile3_touch() {
+        let precompile3 = Address::with_last_byte(3);
+        let other = Address::with_last_byte(4);
+        let mut database = CacheDB::default();
+        database.insert_account_info(&precompile3, AccountInfo::default());
+        database.insert_account_info(&other, AccountInfo::default());
+        let mut state = State::new(database);
+
+        let checkpoint = state.checkpoint();
+        state.touch(&precompile3);
+        state.touch(&other);
+
+        state.rollback(checkpoint, Version::base(SpecId::SPURIOUS_DRAGON).features);
+        assert!(state.touched.contains(&precompile3));
+        assert!(!state.touched.contains(&other));
+    }
+
+    #[test]
+    fn non_revertible_warmth_is_not_journaled_or_rolled_back() {
+        let base_account = Address::with_last_byte(0x10);
+        let frame_account = Address::with_last_byte(0x11);
+        let base_storage = Address::with_last_byte(0x12);
+        let frame_storage = Address::with_last_byte(0x13);
+        let key = Word::from(1);
+        let mut state = State::new(CacheDB::default());
+
+        state.warm_account_non_revertible(&base_account);
+        assert!(state.warm_storage_non_revertible(&base_storage, &key));
+        assert!(state.journal.is_empty());
+
+        let checkpoint = state.checkpoint();
+        assert!(state.warm_account(&frame_account));
+        assert!(state.warm_storage(&frame_storage, &key));
+        assert_eq!(state.journal.len(), 2);
+
+        state.rollback(checkpoint, Version::base(SpecId::FRONTIER).features);
+        assert!(state.is_account_warm(&base_account));
+        assert!(state.is_storage_warm(&base_storage, &key));
+        assert!(!state.is_account_warm(&frame_account));
+        assert!(!state.is_storage_warm(&frame_storage, &key));
+    }
+}
