+++ crates/evm2/src/evm/state/mod.rs
@@ -1,5 +1,17 @@
 //! Basic in-memory EVM host state.
 
+mod account;
+mod changes;
+mod journal;
+mod storage;
+mod tracked;
+
+pub use account::{Account, AccountInfo};
+pub use changes::{StateChanges, StorageChangeSet};
+pub use journal::{JournalEntry, StateCheckpoint};
+pub use storage::StorageOverlay;
+pub use tracked::Tracked;
+
 use super::{
     SStore,
     db::{CacheDB, DatabaseCommit, DbResult, DynDatabase},
@@ -13,332 +25,11 @@ use crate::{
 };
 use alloc::{boxed::Box, collections::BTreeMap, vec::Vec};
 use alloy_primitives::{
-    Address, B256, KECCAK256_EMPTY, Log, U256,
+    Address, B256, KECCAK256_EMPTY, Log,
     map::{AddressMap, AddressSet, U256Map, hash_map},
 };
 use core::mem;
 use derive_where::derive_where;
-
-/// A value tracked together with the value it had at the start of the current
-/// transaction.
-///
-/// `Tracked` is used by [`State`] to keep an overlay over the backing database
-/// and by [`StateChanges`] to describe account and storage transitions.
-/// `original` is the value at the current transaction boundary, while `current`
-/// is the value after all in-flight EVM mutations. When a transaction is
-/// accepted, `current` becomes the next transaction's `original` without writing
-/// anything to the backing database.
-#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
-pub struct Tracked<T> {
-    /// Value at the start of the current transaction.
-    pub original: T,
-    /// Current overlay value.
-    pub current: T,
-    #[doc(hidden)] // Not public API. Please use an existing constructor.
-    pub _non_exhaustive: (),
-}
-
-impl<T> Tracked<T> {
-    /// Creates a tracked value whose original and current values are equal.
-    #[inline]
-    pub fn new(value: T) -> Self
-    where
-        T: Clone,
-    {
-        Self { original: value.clone(), current: value, _non_exhaustive: () }
-    }
-}
-
-impl<T: PartialEq> Tracked<T> {
-    /// Returns whether the current value differs from the original value.
-    #[inline]
-    pub fn is_changed(&self) -> bool {
-        self.original != self.current
-    }
-}
-
-/// Account information loaded from the backing database or emitted in a state
-/// transition.
-#[derive(Clone, Debug, Eq, Hash, PartialEq)]
-pub struct AccountInfo {
-    /// Account balance.
-    pub balance: Word,
-    /// Account nonce.
-    pub nonce: u64,
-    /// Hash of the raw bytes in `code`, or the empty code hash.
-    pub code_hash: B256,
-    /// Bytecode associated with this account.
-    pub code: Option<Bytecode>,
-    #[doc(hidden)] // Not public API. Please use an existing constructor.
-    pub _non_exhaustive: (),
-}
-
-impl Default for AccountInfo {
-    #[inline]
-    fn default() -> Self {
-        Self {
-            balance: U256::ZERO,
-            nonce: 0,
-            code_hash: KECCAK256_EMPTY,
-            code: Some(Bytecode::default()),
-            _non_exhaustive: (),
-        }
-    }
-}
-
-impl AccountInfo {
-    /// Creates a new [`AccountInfo`] with the given fields.
-    #[inline]
-    pub const fn new(balance: Word, nonce: u64, code_hash: B256, code: Bytecode) -> Self {
-        Self { balance, nonce, code_hash, code: Some(code), _non_exhaustive: () }
-    }
-
-    /// Creates a new [`AccountInfo`] with the given code.
-    #[inline]
-    pub fn with_code(self, code: Bytecode) -> Self {
-        Self { code_hash: code.hash_slow(), code: Some(code), ..self }
-    }
-
-    /// Creates a new [`AccountInfo`] with the given balance.
-    #[inline]
-    pub const fn with_balance(mut self, balance: Word) -> Self {
-        self.balance = balance;
-        self
-    }
-
-    /// Creates a new [`AccountInfo`] with the given nonce.
-    #[inline]
-    pub const fn with_nonce(mut self, nonce: u64) -> Self {
-        self.nonce = nonce;
-        self
-    }
-
-    /// Sets account bytecode and updates the code hash.
-    #[inline]
-    pub fn set_code(&mut self, code: Bytecode) {
-        self.code_hash = code.hash_slow();
-        self.code = Some(code);
-    }
-
-    /// Returns whether this account is empty by the Spurious Dragon definition.
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.balance.is_zero() && self.nonce == 0 && self.code_hash == KECCAK256_EMPTY
-    }
-}
-
-/// Mutable account state cached by [`State`].
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
-pub struct Account {
-    /// Account nonce.
-    pub nonce: u64,
-    /// Account balance.
-    pub balance: Word,
-    /// Account code hash.
-    pub code_hash: B256,
-    /// Cached account bytecode.
-    pub code: Bytecode,
-    /// Whether the account was created in the current transaction.
-    pub just_created: bool,
-    /// Whether the account code has been modified.
-    pub code_changed: bool,
-    #[doc(hidden)] // Not public API. Please use an existing constructor.
-    pub _non_exhaustive: (),
-}
-
-impl Account {
-    /// Creates an account from database account info.
-    #[inline]
-    pub fn from_info(info: AccountInfo) -> Self {
-        Self {
-            nonce: info.nonce,
-            balance: info.balance,
-            code_hash: info.code_hash,
-            code: info.code.unwrap_or_default(),
-            just_created: false,
-            code_changed: false,
-            _non_exhaustive: (),
-        }
-    }
-
-    /// Returns account info.
-    #[inline]
-    pub fn info(&self) -> AccountInfo {
-        AccountInfo {
-            balance: self.balance,
-            nonce: self.nonce,
-            code_hash: self.code_hash,
-            code: Some(self.code.clone()),
-            _non_exhaustive: (),
-        }
-    }
-
-    /// Returns whether this account is empty.
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.nonce == 0 && self.balance.is_zero() && self.code_hash == KECCAK256_EMPTY
-    }
-}
-
-/// Persistent storage overlay for one account.
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
-pub struct StorageOverlay {
-    /// Whether consumers must delete all pre-existing storage for the account
-    /// before applying individual slot changes.
-    pub wiped: bool,
-    /// Loaded or changed storage slots.
-    pub slots: U256Map<Tracked<Word>>,
-    #[doc(hidden)] // Not public API. Please use an existing constructor.
-    pub _non_exhaustive: (),
-}
-
-/// Complete state transition and emitted logs produced by a transaction.
-///
-/// `StateChanges` is the public write-set returned in [`crate::TxResult`]. It
-/// is intentionally explicit so embedding clients can update their own database
-/// and compute post-state roots without reimplementing EVM account-lifetime
-/// rules. It also carries the logs emitted by the transaction.
-///
-/// Consumers should apply database changes in this order:
-///
-/// 1. write bytecode from [`Self::code`] for every non-empty code hash they do not already have;
-/// 2. for each [`StorageChangeSet`] whose [`StorageChangeSet::wipe`] flag is true, delete all
-///    storage for that account;
-/// 3. apply each storage slot change: a zero [`Tracked::current`] means delete the slot, otherwise
-///    write the slot value;
-/// 4. apply account changes: `current = Some(..)` means upsert the account, `current = None` means
-///    delete the account.
-///
-/// `evm2` does not write to the backing database. These changes describe what
-/// happened; applying them is the responsibility of the caller.
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
-pub struct StateChanges {
-    /// Account changes keyed by address.
-    ///
-    /// [`Tracked::original`] is the account at the beginning of the transaction.
-    /// [`Tracked::current`] is the account after transaction execution and EVM
-    /// account-lifetime rules have been evaluated. `current = None` is an explicit account
-    /// deletion.
-    pub accounts: BTreeMap<Address, Tracked<Option<AccountInfo>>>,
-    /// Persistent storage changes keyed by account address.
-    ///
-    /// Each slot change's [`Tracked::original`] value is the slot value at the beginning of the
-    /// transaction, after any storage wipe/re-incarnation semantics that occurred before the slot
-    /// was loaded. `current = 0` means the consumer should delete the slot.
-    pub storage: BTreeMap<Address, StorageChangeSet>,
-    /// Newly created or modified bytecode keyed by code hash.
-    pub code: BTreeMap<B256, Bytecode>,
-    /// Logs emitted by the transaction.
-    pub logs: Vec<Log>,
-    #[doc(hidden)] // Not public API. Please use an existing constructor.
-    pub _non_exhaustive: (),
-}
-
-impl StateChanges {
-    /// Returns whether this transition contains no changes or logs.
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.accounts.is_empty()
-            && self.storage.is_empty()
-            && self.code.is_empty()
-            && self.logs.is_empty()
-    }
-}
-
-/// Storage transition for a single account.
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
-pub struct StorageChangeSet {
-    /// If true, delete all pre-existing storage for this account before applying
-    /// [`Self::slots`]. This is used for selfdestruct and contract
-    /// re-incarnation semantics using an explicit storage wipe marker.
-    pub wipe: bool,
-    /// Changed storage slots keyed by slot.
-    pub slots: BTreeMap<Word, Tracked<Word>>,
-    #[doc(hidden)] // Not public API. Please use an existing constructor.
-    pub _non_exhaustive: (),
-}
-
-/// State checkpoint for reverting state changes.
-#[allow(missing_copy_implementations)]
-#[derive(Debug, Eq, PartialEq)]
-pub struct StateCheckpoint {
-    /// Revert journal length at the checkpoint.
-    journal_len: usize,
-    /// Emitted log count at the checkpoint.
-    logs_len: usize,
-}
-
-/// Compact journal entry for reverting state changes.
-#[derive(Clone, Debug, Eq, PartialEq)]
-#[non_exhaustive]
-pub enum JournalEntry {
-    /// Account current value changed.
-    AccountChange {
-        /// Account address.
-        address: Address,
-        /// Previous current account value.
-        previous: Option<Account>,
-    },
-    /// Account overlay entry was inserted.
-    AccountInserted {
-        /// Account address.
-        address: Address,
-    },
-    /// Account was touched.
-    Touch {
-        /// Account address.
-        address: Address,
-    },
-    /// Account was self-destructed.
-    SelfDestruct {
-        /// Account address.
-        address: Address,
-    },
-    /// Persistent storage changed.
-    StorageChange {
-        /// Account address.
-        address: Address,
-        /// Storage key.
-        key: Word,
-        /// Previous current storage value.
-        previous: Word,
-    },
-    /// Persistent storage slot overlay was inserted.
-    StorageInserted {
-        /// Account address.
-        address: Address,
-        /// Storage key.
-        key: Word,
-    },
-    /// Account storage wipe flag changed.
-    StorageWipe {
-        /// Account address.
-        address: Address,
-        /// Previous storage overlay.
-        previous: Option<StorageOverlay>,
-    },
-    /// Transient storage changed.
-    TransientStorageChange {
-        /// Account address.
-        address: Address,
-        /// Storage key.
-        key: Word,
-        /// Previous transient storage value.
-        previous: Option<Word>,
-    },
-    /// Account was warmed by EIP-2929 access tracking.
-    AccountWarmed {
-        /// Account address.
-        address: Address,
-    },
-    /// Storage slot was warmed by EIP-2929 access tracking.
-    StorageWarmed {
-        /// Account address.
-        address: Address,
-        /// Storage key.
-        key: Word,
-    },
-}
 
 /// Mutable EVM state with an accepted-state cache, transaction layer, and reversible journal.
 #[derive_where(Debug)]
@@ -1230,302 +921,5 @@ impl State {
     #[cfg(test)]
     pub(super) fn commit_transaction_overlay(&mut self) {
         let _ = self.accept_transaction();
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{SpecId, constants::EIP7708_BURN_TOPIC, evm::CacheDB};
-    use alloy_primitives::Bytes;
-
-    #[test]
-    fn storage_change_rolls_back_to_checkpoint() {
-        let address = Address::from([0x11; 20]);
-        let mut database = CacheDB::default();
-        database.insert_account_info(&address, AccountInfo::default());
-        database.insert_account_storage(&address, &Word::from(1), &Word::from(10));
-        let mut state = State::new(database);
-
-        let checkpoint = state.checkpoint();
-        state.set_storage(&address, &Word::from(1), &Word::from(20)).unwrap();
-        state.set_storage(&address, &Word::from(1), &Word::from(30)).unwrap();
-
-        assert_eq!(state.storage(&address, &Word::from(1)).unwrap(), Word::from(30));
-        state.rollback(checkpoint, Version::base(SpecId::FRONTIER).features);
-        assert_eq!(state.storage(&address, &Word::from(1)).unwrap(), Word::from(10));
-    }
-
-    #[test]
-    fn transient_storage_change_rolls_back_to_checkpoint() {
-        let address = Address::from([0x22; 20]);
-        let mut state = State::new(CacheDB::default());
-
-        state.set_transient_storage(&address, &Word::from(1), &Word::from(10));
-        let checkpoint = state.checkpoint();
-        state.set_transient_storage(&address, &Word::from(1), &Word::from(20));
-
-        assert_eq!(state.transient_storage(&address, &Word::from(1)), Word::from(20));
-        state.rollback(checkpoint, Version::base(SpecId::FRONTIER).features);
-        assert_eq!(state.transient_storage(&address, &Word::from(1)), Word::from(10));
-    }
-
-    #[test]
-    fn destruct_change_rolls_back_to_checkpoint() {
-        let address = Address::from([0x33; 20]);
-        let mut state = State::new(CacheDB::default());
-
-        let checkpoint = state.checkpoint();
-        state.mark_destructed(&address);
-
-        assert!(state.is_selfdestructed(&address));
-        state.rollback(checkpoint, Version::base(SpecId::FRONTIER).features);
-        assert!(!state.is_selfdestructed(&address));
-    }
-
-    #[test]
-    fn log_rolls_back_to_checkpoint() {
-        use alloy_primitives::{Bytes, LogData};
-
-        let mut state = State::new(CacheDB::default());
-        let kept = Log {
-            address: Address::from([0x44; 20]),
-            data: LogData::new_unchecked(Vec::new(), Bytes::from_static(&[0x01])),
-        };
-        let reverted = Log {
-            address: Address::from([0x55; 20]),
-            data: LogData::new_unchecked(Vec::new(), Bytes::from_static(&[0x02])),
-        };
-
-        state.log(kept.clone());
-        let checkpoint = state.checkpoint();
-        state.log(reverted);
-
-        assert_eq!(
-            state.logs(),
-            &[
-                kept.clone(),
-                Log {
-                    address: Address::from([0x55; 20]),
-                    data: LogData::new_unchecked(Vec::new(), Bytes::from_static(&[0x02])),
-                }
-            ]
-        );
-        state.rollback(checkpoint, Version::base(SpecId::FRONTIER).features);
-        assert_eq!(state.logs(), &[kept]);
-    }
-
-    #[test]
-    fn spurious_dragon_rollback_preserves_precompile3_touch() {
-        let precompile3 = Address::with_last_byte(3);
-        let other = Address::with_last_byte(4);
-        let mut database = CacheDB::default();
-        database.insert_account_info(&precompile3, AccountInfo::default());
-        database.insert_account_info(&other, AccountInfo::default());
-        let mut state = State::new(database);
-
-        let checkpoint = state.checkpoint();
-        state.touch(&precompile3);
-        state.touch(&other);
-
-        state.rollback(checkpoint, Version::base(SpecId::SPURIOUS_DRAGON).features);
-        assert!(state.touched.contains(&precompile3));
-        assert!(!state.touched.contains(&other));
-    }
-
-    #[test]
-    fn non_revertible_warmth_is_not_journaled_or_rolled_back() {
-        let base_account = Address::with_last_byte(0x10);
-        let frame_account = Address::with_last_byte(0x11);
-        let base_storage = Address::with_last_byte(0x12);
-        let frame_storage = Address::with_last_byte(0x13);
-        let key = Word::from(1);
-        let mut state = State::new(CacheDB::default());
-
-        state.warm_account_non_revertible(&base_account);
-        assert!(state.warm_storage_non_revertible(&base_storage, &key));
-        assert!(state.journal.is_empty());
-
-        let checkpoint = state.checkpoint();
-        assert!(state.warm_account(&frame_account));
-        assert!(state.warm_storage(&frame_storage, &key));
-        assert_eq!(state.journal.len(), 2);
-
-        state.rollback(checkpoint, Version::base(SpecId::FRONTIER).features);
-        assert!(state.is_account_warm(&base_account));
-        assert!(state.is_storage_warm(&base_storage, &key));
-        assert!(!state.is_account_warm(&frame_account));
-        assert!(!state.is_storage_warm(&frame_storage, &key));
-    }
-
-    #[test]
-    fn state_changes_take_logs_from_transaction_state() {
-        use alloy_primitives::{Bytes, LogData};
-
-        let mut state = State::new(CacheDB::default());
-        let log = Log {
-            address: Address::from([0x66; 20]),
-            data: LogData::new_unchecked(Vec::new(), Bytes::from_static(&[0x03])),
-        };
-
-        state.log(log.clone());
-        state.finalize_transaction_(Version::base(crate::SpecId::SPURIOUS_DRAGON));
-        let changes = state.build_state_changes();
-        assert_eq!(changes.logs.as_slice(), core::slice::from_ref(&log));
-        assert!(state.logs().is_empty());
-
-        state.commit_transaction_overlay();
-        state.clear_transaction_state();
-        assert!(state.logs().is_empty());
-    }
-
-    #[test]
-    fn transfer_to_self_requires_balance() {
-        let address = Address::from([0x77; 20]);
-        let mut database = CacheDB::default();
-        database.insert_account_info(&address, AccountInfo::default().with_balance(Word::from(3)));
-        let mut state = State::new(database);
-
-        assert!(!state.transfer(&address, &address, &Word::from(4)).unwrap());
-        assert!(state.transfer(&address, &address, &Word::from(3)).unwrap());
-    }
-
-    #[test]
-    fn spurious_dragon_deletes_touched_empty_existing_account() {
-        let address = Address::from([0x44; 20]);
-        let empty = AccountInfo { code: None, ..AccountInfo::default() };
-        let mut database = CacheDB::default();
-        database.insert_account_info(&address, empty.clone());
-        let mut state = State::new(database);
-
-        state.touch(&address);
-        state.finalize_transaction_(Version::base(crate::SpecId::SPURIOUS_DRAGON));
-        let changes = state.build_state_changes();
-
-        let change = changes.accounts.get(&address).expect("touched empty account is deleted");
-        assert_eq!(change.original, Some(empty));
-        assert_eq!(change.current, None);
-        assert!(changes.storage.get(&address).is_some_and(|storage| storage.wipe));
-    }
-
-    #[test]
-    fn homestead_preserves_touched_empty_existing_account() {
-        let address = Address::from([0x45; 20]);
-        let mut database = CacheDB::default();
-        database.insert_account_info(&address, AccountInfo::default());
-        let mut state = State::new(database);
-
-        state.touch(&address);
-        state.finalize_transaction_(Version::base(crate::SpecId::HOMESTEAD));
-        let changes = state.build_state_changes();
-
-        assert!(!changes.accounts.contains_key(&address));
-        assert!(!changes.storage.contains_key(&address));
-    }
-
-    #[test]
-    fn homestead_materializes_touched_empty_new_account() {
-        let address = Address::from([0x46; 20]);
-        let mut state = State::new(CacheDB::default());
-
-        state.touch(&address);
-        state.finalize_transaction_(Version::base(crate::SpecId::HOMESTEAD));
-        let changes = state.build_state_changes();
-
-        let change =
-            changes.accounts.get(&address).expect("pre-spurious empty touch creates account");
-        assert_eq!(change.original, None);
-        let current = change.current.as_ref().expect("created empty account");
-        assert!(current.is_empty());
-    }
-
-    #[test]
-    fn spurious_dragon_ignores_touched_empty_new_account() {
-        let address = Address::from([0x47; 20]);
-        let mut state = State::new(CacheDB::default());
-
-        state.touch(&address);
-        state.finalize_transaction_(Version::base(crate::SpecId::SPURIOUS_DRAGON));
-        let changes = state.build_state_changes();
-
-        assert!(!changes.accounts.contains_key(&address));
-        assert!(!changes.storage.contains_key(&address));
-    }
-
-    #[test]
-    fn finalization_preserves_touched_set_capacity() {
-        let mut state = State::new(CacheDB::default());
-
-        for i in 0..32 {
-            state.touch(&Address::from([i; 20]));
-            state.mark_destructed(&Address::from([i + 32; 20]));
-        }
-
-        let touched_capacity = state.touched.capacity();
-        let selfdestructs_capacity = state.selfdestructs.capacity();
-
-        state.finalize_transaction_(Version::base(crate::SpecId::SPURIOUS_DRAGON));
-
-        assert!(state.touched.is_empty());
-        assert!(state.selfdestructs.is_empty());
-        assert_eq!(state.touched.capacity(), touched_capacity);
-        assert_eq!(state.selfdestructs.capacity(), selfdestructs_capacity);
-    }
-
-    #[test]
-    fn selfdestruct_deletes_account_and_wipes_storage() {
-        let address = Address::from([0x48; 20]);
-        let mut database = CacheDB::default();
-        database.insert_account_info(&address, AccountInfo::default().with_balance(Word::from(1)));
-        database.insert_account_storage(&address, &Word::from(1), &Word::from(2));
-        let mut state = State::new(database);
-
-        state.mark_destructed(&address);
-        state.finalize_transaction_(Version::base(crate::SpecId::SPURIOUS_DRAGON));
-        let changes = state.build_state_changes();
-
-        let change = changes.accounts.get(&address).expect("selfdestruct deletes account");
-        assert!(change.original.is_some());
-        assert_eq!(change.current, None);
-        assert!(changes.storage.get(&address).is_some_and(|storage| storage.wipe));
-    }
-
-    #[test]
-    fn eip7708_delayed_burn_logs_selfdestructs_sorted() {
-        let high = Address::from([0x22; 20]);
-        let low = Address::from([0x11; 20]);
-        let mut database = CacheDB::default();
-        database.insert_account_info(&high, AccountInfo::default().with_balance(Word::from(2)));
-        database.insert_account_info(&low, AccountInfo::default().with_balance(Word::from(1)));
-        let mut state = State::new(database);
-
-        state.mark_destructed(&high);
-        state.mark_destructed(&low);
-        let mut inspected = Vec::new();
-        state
-            .finalize_transaction(Version::base(SpecId::AMSTERDAM), |log| {
-                inspected.push(log.clone())
-            })
-            .unwrap();
-
-        let changes = state.build_state_changes();
-        assert_eq!(inspected, changes.logs);
-        assert_eq!(changes.logs.len(), 2);
-        assert_eq!(
-            changes.logs[0].topics(),
-            &[EIP7708_BURN_TOPIC, B256::left_padding_from(low.as_slice())]
-        );
-        assert_eq!(
-            changes.logs[0].data.data,
-            Bytes::copy_from_slice(&Word::from(1).to_be_bytes::<32>())
-        );
-        assert_eq!(
-            changes.logs[1].topics(),
-            &[EIP7708_BURN_TOPIC, B256::left_padding_from(high.as_slice())]
-        );
-        assert_eq!(
-            changes.logs[1].data.data,
-            Bytes::copy_from_slice(&Word::from(2).to_be_bytes::<32>())
-        );
     }
 }
